Breaks pipeline into sub modules
